% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{AU Calculator user interface}
\usage{
app_ui()
}
\value{
A \code{\link[shiny:reexports]{shiny::tagList()}} containing the app UI definition.
}
\description{
AU Calculator user interface
}
