% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{dataInputUI}
\alias{dataInputUI}
\title{Data upload and column selection UI}
\usage{
dataInputUI(id)
}
\arguments{
\item{id}{Module namespace identifier.}
}
\value{
A list of UI elements that render the file upload tools.
}
\description{
Data upload and column selection UI
}
