% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_params.R
\name{parameterServer}
\alias{parameterServer}
\title{Parameter server logic}
\usage{
parameterServer(id)
}
\arguments{
\item{id}{Module namespace identifier.}
}
\value{
A list with reactivity for acreage, intake, and button presses.
}
\description{
Parameter server logic
}
