% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_params.R
\name{parameterUI}
\alias{parameterUI}
\title{Parameter inputs UI}
\usage{
parameterUI(id)
}
\arguments{
\item{id}{Module namespace identifier.}
}
\value{
UI controls for the acreage, calculate button, and intake choice.
}
\description{
Parameter inputs UI
}
