% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{resultsMainUI}
\alias{resultsMainUI}
\title{Main results panel}
\usage{
resultsMainUI(id)
}
\arguments{
\item{id}{Module namespace identifier.}
}
\value{
The UI layout used in the main panel (AU value, download, formula).
}
\description{
Main results panel
}
