% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{resultsServer}
\alias{resultsServer}
\title{Results server logic}
\usage{
resultsServer(id, data_inputs, parameter_inputs, k_const)
}
\arguments{
\item{id}{Module namespace identifier.}

\item{data_inputs}{Reactive list returned from \code{dataInputServer()}.}

\item{parameter_inputs}{Reactive list returned from \code{parameterServer()}.}

\item{k_const}{Numeric conversion factor.}
}
\description{
Results server logic
}
