% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{resultsSidebarUI}
\alias{resultsSidebarUI}
\title{Sidebar summary output}
\usage{
resultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace identifier.}
}
\value{
A verbatim text output that can be dropped into the sidebar.
}
\description{
Sidebar summary output
}
