% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{server}
\alias{server}
\alias{app_server}
\title{Shiny App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\description{
The server logic for the Animal Units (AUs) Grazing Calculator Shiny app.
}
