% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{tableServer}
\alias{tableServer}
\title{Data table server logic}
\usage{
tableServer(id, data_inputs)
}
\arguments{
\item{id}{Module namespace identifier.}

\item{data_inputs}{Reactive list returned from \code{dataInputServer()}.}
}
\description{
Data table server logic
}
